Fized player name realloc size and attempted to create a finite buffer size solution

+++ identifiers.h
@@ -9,6 +9,6 @@ typedef enum { false, true } bool;
 #define PLAYER_2_TOKEN 2
 #define EMPTY_SLOT 0
 
-#define ARRAY_LENGTH(x) ((int)sizeof(x) / sizeof((x)[0])) //we use this because we cannot calculate the size of a dynamic array; the compiler can never know its size
+#define ARRAY_LENGTH(x) ((int)sizeof(x) / (int)sizeof((x)[0])) //we use this because we cannot calculate the size of a dynamic array; the compiler can never know its size
 
 #define NUL '\0' //used to "nullify" a char
